Move gui utilities to own folder
