+++ src/gui/content_widget/content_widget.cpp
@@ -1,6 +1,6 @@
 #include "content_widget/content_widget.h"
 
-#include "gui_utility.h"
+#include "gui_utils/graphics.h"
 #include "hal_content_anchor/hal_content_anchor.h"
 #include "toolbar/toolbar.h"
