+++ src/gui/gui_utils/sort.cpp
@@ -1,4 +1,4 @@
-#include "gui_utility/sort.h"
+#include "gui_utils/sort.h"
 
 #include "def.h"
