+++ src/gui/welcome_screen/open_file_widget.cpp
@@ -1,7 +1,7 @@
 #include "welcome_screen/open_file_widget.h"
 
 #include "file_manager/file_manager.h"
-#include "gui_utility.h"
+#include "gui_utils/graphics.h"
 
 #include <QDragEnterEvent>
 #include <QLabel>
